Rename commands
Took 19 minutes
