+++ DentalCore.Wpf/Commands/AsyncRelayCommand.cs
@@ -1,19 +1,20 @@
 ﻿using System;
 using System.Threading.Tasks;
+using DentalCore.Wpf.Abstract;
 
 namespace DentalCore.Wpf.Commands;
 
-public class AsyncRelayCommand<T> : BaseAsyncCommand
+public class AsyncRelayCommand : BaseAsyncCommand
 {
-    private readonly Func<T, Task> _execute;
+    private readonly Func<Task> _execute;
 
-    public AsyncRelayCommand(Func<T, Task> execute, Action<Exception>? onException = null) : base(onException)
+    public AsyncRelayCommand(Func<Task> execute, Action<Exception>? onException = null) : base(onException)
     {
         _execute = execute;
     }
 
     protected override async Task ExecuteAsync(object? parameter)
     {
-        await _execute((T)parameter!);
+        await _execute();
     }
 }
