+++ DentalCore.Wpf/Commands/Generic/RelayCommand.cs
@@ -1,21 +1,21 @@
 ﻿using System;
 using System.Windows.Input;
 
-namespace DentalCore.Wpf.Commands;
+namespace DentalCore.Wpf.Commands.Generic;
 
-public class RelayCommand : ICommand
+public class RelayCommand<T> : ICommand
 {
-    private readonly Action _execute;
-    private readonly Func<bool>? _canExecute;
+    private readonly Action<T> _execute;
+    private readonly Predicate<T>? _canExecute;
     
     public bool CanExecute(object? parameter)
     {
-        return _canExecute is null || _canExecute();
+        return _canExecute is null || _canExecute((T)parameter!);
     }
 
     public void Execute(object? parameter)
     {
-        _execute();
+        _execute((T)parameter!);
     }
 
     public event EventHandler? CanExecuteChanged {
@@ -23,7 +23,7 @@ public class RelayCommand : ICommand
         remove => CommandManager.RequerySuggested -= value;
     }
 
-    public RelayCommand(Action execute, Func<bool>? canExecute = null)
+    public RelayCommand(Action<T> execute, Predicate<T>? canExecute = null)
     {
         _execute = execute ?? throw new ArgumentNullException(nameof(execute));
         _canExecute = canExecute;
